[Design] SelectView 레이아웃 조정
